chore: refactor api test helper
