+++ internal/orgs/list/list_test.go
@@ -18,7 +18,7 @@ func TestOrganizationListCommand(t *testing.T) {
 		token := apitest.RandomAccessToken(t)
 		t.Setenv("SUPABASE_ACCESS_TOKEN", string(token))
 		// Flush pending mocks after test execution
-		defer gock.Off()
+		defer gock.OffAll()
 		gock.New("https://api.supabase.io").
 			Get("/v1/organizations").
 			Reply(200).
@@ -30,6 +30,8 @@ func TestOrganizationListCommand(t *testing.T) {
 			})
 		// Run test
 		assert.NoError(t, Run(fsys))
+		// Validate api
+		assert.Empty(t, apitest.ListUnmatchedRequests())
 	})
 
 	t.Run("throws error on failure to load token", func(t *testing.T) {
@@ -43,12 +45,14 @@ func TestOrganizationListCommand(t *testing.T) {
 		token := apitest.RandomAccessToken(t)
 		t.Setenv("SUPABASE_ACCESS_TOKEN", string(token))
 		// Flush pending mocks after test execution
-		defer gock.Off()
+		defer gock.OffAll()
 		gock.New("https://api.supabase.io").
 			Get("/v1/organizations").
 			ReplyError(errors.New("network error"))
 		// Run test
 		assert.Error(t, Run(fsys))
+		// Validate api
+		assert.Empty(t, apitest.ListUnmatchedRequests())
 	})
 
 	t.Run("throws error on server unavailable", func(t *testing.T) {
@@ -58,13 +62,15 @@ func TestOrganizationListCommand(t *testing.T) {
 		token := apitest.RandomAccessToken(t)
 		t.Setenv("SUPABASE_ACCESS_TOKEN", string(token))
 		// Flush pending mocks after test execution
-		defer gock.Off()
+		defer gock.OffAll()
 		gock.New("https://api.supabase.io").
 			Get("/v1/organizations").
 			Reply(500).
 			JSON(map[string]string{"message": "unavailable"})
 		// Run test
 		assert.Error(t, Run(fsys))
+		// Validate api
+		assert.Empty(t, apitest.ListUnmatchedRequests())
 	})
 
 	t.Run("throws error on malformed json", func(t *testing.T) {
@@ -74,12 +80,14 @@ func TestOrganizationListCommand(t *testing.T) {
 		token := apitest.RandomAccessToken(t)
 		t.Setenv("SUPABASE_ACCESS_TOKEN", string(token))
 		// Flush pending mocks after test execution
-		defer gock.Off()
+		defer gock.OffAll()
 		gock.New("https://api.supabase.io").
 			Get("/v1/organizations").
 			Reply(200).
 			JSON(map[string]string{})
 		// Run test
 		assert.Error(t, Run(fsys))
+		// Validate api
+		assert.Empty(t, apitest.ListUnmatchedRequests())
 	})
 }
