+++ internal/secrets/list/list_test.go
@@ -28,7 +28,7 @@ func TestSecretListCommand(t *testing.T) {
 		token := apitest.RandomAccessToken(t)
 		t.Setenv("SUPABASE_ACCESS_TOKEN", string(token))
 		// Flush pending mocks after test execution
-		defer gock.Off()
+		defer gock.OffAll()
 		gock.New("https://api.supabase.io").
 			Get("/v1/projects/" + project + "/secrets").
 			Reply(200).
@@ -40,6 +40,8 @@ func TestSecretListCommand(t *testing.T) {
 			})
 		// Run test
 		assert.NoError(t, Run(context.Background(), fsys))
+		// Validate api
+		assert.Empty(t, apitest.ListUnmatchedRequests())
 	})
 
 	t.Run("throws error on missing config file", func(t *testing.T) {
@@ -79,12 +81,14 @@ func TestSecretListCommand(t *testing.T) {
 		token := apitest.RandomAccessToken(t)
 		t.Setenv("SUPABASE_ACCESS_TOKEN", string(token))
 		// Flush pending mocks after test execution
-		defer gock.Off()
+		defer gock.OffAll()
 		gock.New("https://api.supabase.io").
 			Get("/v1/projects/" + project + "/secrets").
 			ReplyError(errors.New("network error"))
 		// Run test
 		assert.Error(t, Run(context.Background(), fsys))
+		// Validate api
+		assert.Empty(t, apitest.ListUnmatchedRequests())
 	})
 
 	t.Run("throws error on server unavailable", func(t *testing.T) {
@@ -100,13 +104,15 @@ func TestSecretListCommand(t *testing.T) {
 		token := apitest.RandomAccessToken(t)
 		t.Setenv("SUPABASE_ACCESS_TOKEN", string(token))
 		// Flush pending mocks after test execution
-		defer gock.Off()
+		defer gock.OffAll()
 		gock.New("https://api.supabase.io").
 			Get("/v1/projects/" + project + "/secrets").
 			Reply(500).
 			JSON(map[string]string{"message": "unavailable"})
 		// Run test
 		assert.Error(t, Run(context.Background(), fsys))
+		// Validate api
+		assert.Empty(t, apitest.ListUnmatchedRequests())
 	})
 
 	t.Run("throws error on malformed json", func(t *testing.T) {
@@ -122,12 +128,14 @@ func TestSecretListCommand(t *testing.T) {
 		token := apitest.RandomAccessToken(t)
 		t.Setenv("SUPABASE_ACCESS_TOKEN", string(token))
 		// Flush pending mocks after test execution
-		defer gock.Off()
+		defer gock.OffAll()
 		gock.New("https://api.supabase.io").
 			Get("/v1/projects/" + project + "/secrets").
 			Reply(200).
 			JSON(map[string]string{})
 		// Run test
 		assert.Error(t, Run(context.Background(), fsys))
+		// Validate api
+		assert.Empty(t, apitest.ListUnmatchedRequests())
 	})
 }
