+++ internal/secrets/unset/unset_test.go
@@ -28,7 +28,7 @@ func TestSecretUnsetCommand(t *testing.T) {
 		token := apitest.RandomAccessToken(t)
 		t.Setenv("SUPABASE_ACCESS_TOKEN", string(token))
 		// Flush pending mocks after test execution
-		defer gock.Off()
+		defer gock.OffAll()
 		gock.New("https://api.supabase.io").
 			Delete("/v1/projects/" + project + "/secrets").
 			MatchType("json").
@@ -36,6 +36,8 @@ func TestSecretUnsetCommand(t *testing.T) {
 			Reply(200)
 		// Run test
 		assert.NoError(t, Run(context.Background(), []string{"my-secret"}, fsys))
+		// Validate api
+		assert.Empty(t, apitest.ListUnmatchedRequests())
 	})
 
 	t.Run("throws error on missing config file", func(t *testing.T) {
@@ -91,14 +93,16 @@ func TestSecretUnsetCommand(t *testing.T) {
 		token := apitest.RandomAccessToken(t)
 		t.Setenv("SUPABASE_ACCESS_TOKEN", string(token))
 		// Flush pending mocks after test execution
-		defer gock.Off()
+		defer gock.OffAll()
 		gock.New("https://api.supabase.io").
-			Post("/v1/projects/" + project + "/secrets").
+			Delete("/v1/projects/" + project + "/secrets").
 			MatchType("json").
 			JSON(api.DeleteSecretsJSONBody{"my-secret"}).
 			ReplyError(errors.New("network error"))
 		// Run test
 		assert.Error(t, Run(context.Background(), []string{"my-secret"}, fsys))
+		// Validate api
+		assert.Empty(t, apitest.ListUnmatchedRequests())
 	})
 
 	t.Run("throws error on server unavailable", func(t *testing.T) {
@@ -114,14 +118,16 @@ func TestSecretUnsetCommand(t *testing.T) {
 		token := apitest.RandomAccessToken(t)
 		t.Setenv("SUPABASE_ACCESS_TOKEN", string(token))
 		// Flush pending mocks after test execution
-		defer gock.Off()
+		defer gock.OffAll()
 		gock.New("https://api.supabase.io").
-			Post("/v1/projects/" + project + "/secrets").
+			Delete("/v1/projects/" + project + "/secrets").
 			MatchType("json").
 			JSON(api.DeleteSecretsJSONBody{"my-secret"}).
 			Reply(500).
 			JSON(map[string]string{"message": "unavailable"})
 		// Run test
 		assert.Error(t, Run(context.Background(), []string{"my-secret"}, fsys))
+		// Validate api
+		assert.Empty(t, apitest.ListUnmatchedRequests())
 	})
 }
